+++ reactive-http-client-starter/src/main/java/io/github/huynhngochuyhoang/httpstarter/config/ReactiveHttpClientRuntimeHints.java
@@ -23,6 +23,7 @@ public class ReactiveHttpClientRuntimeHints implements RuntimeHintsRegistrar {
             PathVar.class,
             QueryParam.class,
             HeaderParam.class,
+            IdempotencyKey.class,
             Body.class,
             MultipartBody.class,
             FormField.class,